requirements.txt
botbuilder-core
botbuilder-ai
aiohttp
python-dotenv